LibGfx/ImageFormats: Move BilevelImage to its own file
Pure code move, no behavior change.
